+++ src/import_common/normalization.py
@@ -20,6 +20,16 @@ def remove_extra_spaces(s: str) -> str:
     return s
 
 
+def normalize_first_name(s: str) -> str | None:
+    if not isinstance(s, str) or s is None or s == "" or not s:
+        return
+    return remove_extra_spaces(s)
+
+
+normalize_last_name = normalize_first_name
+normalize_publisher = normalize_first_name
+
+
 def normalize_nullboleanfield(s: Union[str, None, bool]) -> Union[bool, None]:
     if isinstance(s, bool):
         return s
@@ -87,8 +97,8 @@ def normalize_nazwa_dyscypliny(s):
     return normalize_skrot(s)
 
 
-def normalize_isbn(isbn):
-    if isbn is None:
+def normalize_isbn(isbn: str) -> str | None:
+    if isbn is None or not isinstance(isbn, str) or not isbn:
         return
 
     return isbn.replace(".", "").replace("-", "").replace(" ", "").strip()
@@ -114,15 +124,25 @@ def normalize_public_uri(public_uri):
         return public_uri.strip()
 
 
-def normalize_doi(s):
+def normalize_doi(s: str) -> None | str:
+    """
+    https://www.doi.org/doi_handbook/2_Numbering.html#2.4
+    """
+
     if s is None:
-        return s
+        return
+
+    s = s.strip()
+
+    if not s:
+        return
+
     return (
-        s.strip()
+        s.lower()
         .replace("http://", "")
         .replace("https://", "")
+        .replace("dx.doi.org/", "")
         .replace("doi.org/", "")
-        .lower()
     )
 
 
@@ -132,3 +152,14 @@ def normalize_filename(s: str) -> str:
     while s.find("__") != -1:
         s = s.replace("__", "_")
     return s
+
+
+def normalize_orcid(s: str) -> str | None:
+    if not isinstance(s, str) or s is None or s == "" or not s:
+        return
+    return (
+        s.strip()
+        .lower()
+        .replace("http://orcid.org/", "")
+        .replace("https://orcid.org/", "")
+    )